+++ admin/includes/view_all_posts.php
@@ -182,12 +182,18 @@
 
     if(isset($_GET['delete']))
     {
-        $the_post_id = $_GET['delete'];
-        
-        $query = "DELETE FROM posts WHERE post_id = {$the_post_id}";
-        $delete_query = mysqli_query($connection, $query);
-        
-        header("Location: posts.php");
+        if(isset($_SESSION['user_role']))
+        {
+            if($_SESSION['user_role'] == 'admin')
+            {
+                $the_post_id = mysqli_real_escape_string($connection, $_GET['delete']);
+
+                $query = "DELETE FROM posts WHERE post_id = {$the_post_id}";
+                $delete_query = mysqli_query($connection, $query);
+
+                header("Location: posts.php");
+            }
+        }
     }
 
     if(isset($_GET['reset']))
