+++ admin/includes/view_all_users.php
@@ -51,41 +51,46 @@
 </div>
 
 <?php
+if(isset($_GET['change_to_admin']))
+{
+    $the_user_id = $_GET['change_to_admin'];
 
-    if(isset($_GET['change_to_admin']))
+    $query = "UPDATE users SET user_role = 'admin' WHERE user_id = $the_user_id";
+    $admin_query = mysqli_query($connection, $query);
+
+    confirm($admin_query);
+
+    header("Location: users.php");
+}
+
+if(isset($_GET['change_to_sub']))
+{
+    $the_user_id = $_GET['change_to_sub'];
+
+    $query = "UPDATE users SET user_role = 'subscriber' WHERE user_id = $the_user_id";
+    $subscriber_query = mysqli_query($connection, $query);
+
+    confirm($subscriber_query);
+
+    header("Location: users.php");
+}
+
+if(isset($_GET['delete']))
+{
+    if(isset($_SESSION['user_role']))
     {
-        $the_user_id = $_GET['change_to_admin'];
-        
-        $query = "UPDATE users SET user_role = 'admin' WHERE user_id = $the_user_id";
-        $admin_query = mysqli_query($connection, $query);
-        
-        confirm($admin_query);
-        
-        header("Location: users.php");
-    }
+        if($_SESSION['user_role'] == 'admin')
+        {
+            $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
 
-    if(isset($_GET['change_to_sub']))
-    {
-        $the_user_id = $_GET['change_to_sub'];
-        
-        $query = "UPDATE users SET user_role = 'subscriber' WHERE user_id = $the_user_id";
-        $subscriber_query = mysqli_query($connection, $query);
-        
-        confirm($subscriber_query);
-        
-        header("Location: users.php");
-    }
+            $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+            $delete_query = mysqli_query($connection, $query);
 
-    if(isset($_GET['delete']))
-    {
-        $the_user_id = $_GET['delete'];
-        
-        $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-        $delete_query = mysqli_query($connection, $query);
-    
-        confirm($delete_query);
-        
-        header("Location: users.php");
-    }
+            confirm($delete_query);
 
+            header("Location: users.php");
+
+        }
+    }
+}
 ?>
